+++ tests/testthat/test_apply_EfficiencyCorrection.R
@@ -3,14 +3,21 @@ set.seed(1)
 data(ExampleData.XSYG, envir = environment())
 eff_data <- data.frame(WAVELENGTH = 1:1000, runif(1000))
 
+## store in tempfile to import
+eff_data_file <- tempfile(fileext = ".csv")
+write.csv(file = eff_data_file, x = eff_data, row.names = FALSE)
+
 test_that("input validation", {
   testthat::skip_on_cran()
 
   expect_error(apply_EfficiencyCorrection(object = "error"),
                "'object' should be of class 'RLum.Data.Spectrum'")
   expect_error(apply_EfficiencyCorrection(TL.Spectrum,
-                                          spectral.efficiency = "error"),
+                                          spectral.efficiency = 99),
                "'spectral.efficiency' should be of class 'data.frame'")
+  expect_error(apply_EfficiencyCorrection(TL.Spectrum,
+                                          spectral.efficiency = "error"),
+               "`spectral.efficiency` does not provide a valid file path!")
   expect_error(apply_EfficiencyCorrection(TL.Spectrum,
                                           spectral.efficiency = data.frame()),
                "'spectral.efficiency' cannot be an empty data.frame")
@@ -53,6 +60,12 @@ test_that("check functionality", {
   input <- list(a = "test", TL.Spectrum,set_RLum("RLum.Analysis", records = list(TL.Spectrum)))
   expect_warning(apply_EfficiencyCorrection(input, eff_data),
                  "Skipping 'character' object in input list")
+  
+  ## with file input
+  expect_s4_class(
+  apply_EfficiencyCorrection(
+    set_RLum("RLum.Analysis",records = list(TL.Spectrum)), spectral.efficiency = eff_data_file), 
+  "RLum.Analysis")
 })
 
 test_that("snapshot tests", {